ai-engine/requirements.txt
# Deep Learning e Computer Vision
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.25.2
scipy==1.11.4

# Modelos pré-treinados e transfer learning
transformers==4.36.2
timm==0.9.12
ultralytics==8.0.196

# Processamento de vídeo
imageio==2.31.6
imageio-ffmpeg==0.4.9
scikit-video==1.1.11

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
scikit-image==0.22.0

# MLOps e experimentos
mlflow==2.8.1
wandb==0.16.0
tensorboard==2.15.1

# Processamento de dados
kafka-python==2.0.2
redis==5.0.1
pymongo==4.6.0
psycopg2-binary==2.9.9

# APIs e comunicação
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
websockets==12.0

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
structlog==23.2.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0

# Otimização e performance
numba==0.58.1
cython==3.0.6

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
